Wave 2+3: Backend handlers + Tensor/NN IR migration
Wave 2.1 — CpuBackend compile + dispatch handlers:
- UpsampleNearest2d, UpsampleBilinear2d (nearest-neighbor/bilinear)
- AdaptiveAvgPool2d (adaptive average pooling)
- Repeat, CumSum, Erf, Flip, Where (element-wise conditional)

Wave 2.2 — WgpuBackend:
- Erf added to element-wise WGSL shader (opcode 36)
- All other ops registered as CPU fallback via UnsupportedOp

Wave 3 — Tensor/NN method migration to exec_aot():
- indexing.rs: cat, gather, repeat, where_tensor → exec_aot
- reductions.rs: cumsum → exec_aot
- ops.rs: erf → exec_aot
- shape.rs: flip → exec_aot (with slice-based dims API)
- nn/upsample.rs: nearest/bilinear → exec_aot
- nn/activations.rs: AdaptiveAvgPool2d → exec_aot

+++ src/python/nn.rs
@@ -873,7 +873,7 @@ impl_nn_module!(AdaptiveAvgPool2d {
     #[new]
     fn new(output_h: i64, output_w: i64) -> Self {
         AdaptiveAvgPool2d {
-            inner: core_nn::activations::AdaptiveAvgPool2d::new((output_h, output_w)),
+            inner: core_nn::activations::AdaptiveAvgPool2d::new(output_h as usize, output_w as usize),
         }
     }
 });
